requirements.txt
appdirs==1.4.4
asgiref==3.5.2
autopep8==1.6.0
certifi==2024.7.4
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.25
django-widget-tweaks==1.4.12
djangorestframework==3.15.2
idna==3.7
pycodestyle==2.8.0
pycryptodome==3.19.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
urllib3==1.26.19
wheel==0.42.0
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
